Add files via upload
KAUST Competition on Spatial Statistics for Large Datasets
Introduction 
With the development of observing techniques and computing devices, it has become easier and more common to obtain large datasets. Statistical inference in spatial statistics becomes computationally challenging. For decades, various approximation methods have been proposed to model and analyze large-scale spatial data when the exact computation is infeasible. However, in the literature, the performance of the statistical inference using those proposed approximation methods was usually assessed with small and medium datasets only, for which the exact solution can be obtained. Then, for real-world large datasets, the exact computation was no longer feasible. The inference with approximation methods was often validated empirically or via prediction accuracy with the fitted model.

In this competition, the goal is to reassess existing approximation methods on large spatial datasets in a uniform way that guarantees a fair comparison. 
https://cemse.kaust.edu.sa/stsds/kaust-competition-spatial-statistics-large-datasets